Canon EOS trigger bugfix.

+++ src/core/canon_eos1000d.h
@@ -14,7 +14,7 @@ class CanonEOS1000D : public CameraController {
             shoot_internal(duration_seconds * 1000, delay_seconds * 1000);
         }
 
-        void shoot_internal(int duration_ms, int delay_ms) {
+        void shoot_internal(unsigned long duration_ms, unsigned long delay_ms) {
             digitalWrite(TRIGGER_PIN, HIGH);
             _last_delay = delay_ms ;
             _last_duration = duration_ms ;
